O3-5064: Simplify session registry to core functionality
- Removed: SessionMetadata, scheduled cleanup, new REST endpoints
- Kept: SessionLease check-out pattern, per-session locking, auto-start registration
- Tests: 16/16 passing
