+++ api/src/main/java/org/openmrs/module/drools/session/SessionRegistry.java
@@ -2,37 +2,25 @@ package org.openmrs.module.drools.session;
 
 import org.kie.api.runtime.KieSession;
 
-import java.util.Collection;
-import java.util.Optional;
 import java.util.concurrent.TimeUnit;
 
 /**
- * Thread-safe registry for managing stateful KieSession instances.
- * Supports session lifecycle management, metadata tracking, and cleanup operations.
+ * Registry for managing stateful KieSession instances with thread-safe access.
  * 
  * <p>This registry uses a check-out/return pattern to prevent concurrent access issues.
  * Sessions should be checked out using {@link #checkOutSession(String, long, TimeUnit)},
  * used within a try-with-resources block, and automatically returned when the lease closes.</p>
  */
-public interface ThreadSafeSessionRegistry {
+public interface SessionRegistry {
 	
 	/**
-	 * Registers a KieSession in the registry with metadata.
+	 * Registers a KieSession in the registry.
 	 * 
 	 * @param sessionId the unique identifier for the session
 	 * @param session the KieSession instance to register
-	 * @param autoStartable whether the session is auto-startable (started at module initialization)
 	 * @return true if registration was successful, false if session already exists
 	 */
-	boolean registerSession(String sessionId, KieSession session, boolean autoStartable);
-	
-	/**
-	 * Retrieves a KieSession by its identifier.
-	 * 
-	 * @param sessionId the unique identifier for the session
-	 * @return an Optional containing the session if found, empty otherwise
-	 */
-	Optional<KieSession> getSession(String sessionId);
+	boolean registerSession(String sessionId, KieSession session);
 	
 	/**
 	 * Checks out a session for exclusive use, blocking if necessary until the lock is available.
@@ -61,28 +49,6 @@ public interface ThreadSafeSessionRegistry {
 			throws InterruptedException, java.util.concurrent.TimeoutException;
 	
 	/**
-	 * Removes a session from the registry and disposes it.
-	 * 
-	 * @param sessionId the unique identifier for the session to remove
-	 * @return true if the session was removed, false if it didn't exist
-	 */
-	boolean removeSession(String sessionId);
-	
-	/**
-	 * Gets all active session IDs currently in the registry.
-	 * 
-	 * @return a collection of all active session identifiers
-	 */
-	Collection<String> getActiveSessionIds();
-	
-	/**
-	 * Gets session IDs for auto-startable sessions only.
-	 * 
-	 * @return a collection of auto-startable session identifiers
-	 */
-	Collection<String> getAutoStartableSessionIds();
-	
-	/**
 	 * Checks if a session exists in the registry.
 	 * 
 	 * @param sessionId the unique identifier to check
@@ -91,23 +57,8 @@ public interface ThreadSafeSessionRegistry {
 	boolean sessionExists(String sessionId);
 	
 	/**
-	 * Retrieves metadata for a specific session.
-	 * 
-	 * @param sessionId the unique identifier for the session
-	 * @return the session metadata, or null if session doesn't exist
-	 */
-	SessionMetadata getSessionMetadata(String sessionId);
-	
-	/**
 	 * Cleans up expired or stale sessions based on implementation-specific criteria.
-	 * This method is intended to be called periodically or on-demand.
+	 * This method can be called manually when needed.
 	 */
 	void cleanupExpiredSessions();
-	
-	/**
-	 * Gets the total count of active sessions in the registry.
-	 * 
-	 * @return the number of active sessions
-	 */
-	long getActiveSessionCount();
 }
