+++ api/src/main/resources/moduleApplicationContext.xml
@@ -10,7 +10,6 @@
        xmlns:tx="http://www.springframework.org/schema/tx"
        xmlns:aop="http://www.springframework.org/schema/aop"
        xmlns:util="http://www.springframework.org/schema/util"
-       xmlns:task="http://www.springframework.org/schema/task"
        xsi:schemaLocation="http://www.springframework.org/schema/beans
   		    http://www.springframework.org/schema/beans/spring-beans-3.0.xsd
   		    http://www.springframework.org/schema/context
@@ -22,14 +21,9 @@
   		    http://www.springframework.org/schema/aop
   		    http://www.springframework.org/schema/aop/spring-aop-3.0.xsd
   		    http://www.springframework.org/schema/util
-  		    http://www.springframework.org/schema/util/spring-util-3.0.xsd
-  		    http://www.springframework.org/schema/task
-  		    http://www.springframework.org/schema/task/spring-task-3.0.xsd">
+  		    http://www.springframework.org/schema/util/spring-util-3.0.xsd">
 
     <context:component-scan base-package="org.openmrs.module.drools" />
-    
-    <!-- Enable Spring @Scheduled annotation support for automated cleanup -->
-    <task:annotation-driven />
 
    <bean name="droolsEngineService" class="org.openmrs.module.drools.api.impl.DroolsEngineServiceImpl"/>
