+++ omod/src/main/java/org/openmrs/module/drools/web/controller/DroolsSessionController.java
@@ -4,8 +4,6 @@ import org.apache.commons.lang3.ObjectUtils;
 import org.openmrs.module.drools.api.DroolsEngineService;
 import org.openmrs.module.drools.param.DroolsParameterDefinition;
 import org.openmrs.module.drools.session.DroolsSessionConfig;
-import org.openmrs.module.drools.session.SessionMetadata;
-import org.openmrs.module.drools.session.ThreadSafeSessionRegistry;
 import org.openmrs.module.drools.web.DroolsSessionExecutor;
 import org.openmrs.module.webservices.rest.SimpleObject;
 import org.openmrs.module.webservices.rest.web.RestConstants;
@@ -21,7 +19,6 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Collection;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
@@ -39,9 +36,6 @@ public class DroolsSessionController extends BaseRestController {
 
     @Autowired
     private DroolsSessionExecutor sessionExecutor;
-
-    @Autowired
-    private ThreadSafeSessionRegistry sessionRegistry;
 
 
 
@@ -69,133 +63,6 @@ public class DroolsSessionController extends BaseRestController {
         validateParams(sessionConfig, allParams);
         // FIX: Remove duplicate executeSession call - execute only once
         return convertToSimpleObject(sessionExecutor.executeSession(sessionId, allParams), request);
-    }
-
-    /**
-     * GET /ws/rest/v1/drools/rules
-     * Retrieves information about all active sessions in the registry.
-     *
-     * @return SimpleObject containing activeIds, autoStartableIds, and totalCount
-     */
-    @RequestMapping(value = "/rules", method = RequestMethod.GET)
-    @ResponseBody
-    public SimpleObject getActiveSessions() {
-        Collection<String> activeIds = sessionRegistry.getActiveSessionIds();
-        Collection<String> autoStartableIds = sessionRegistry.getAutoStartableSessionIds();
-        long totalCount = sessionRegistry.getActiveSessionCount();
-
-        SimpleObject result = new SimpleObject();
-        result.add("activeSessionIds", activeIds);
-        result.add("autoStartableSessionIds", autoStartableIds);
-        result.add("totalCount", totalCount);
-        return result;
-    }
-
-    /**
-     * GET /ws/rest/v1/drools/rules/{sessionId}
-     * Retrieves detailed metadata for a specific session.
-     *
-     * @param sessionId the session identifier
-     * @return SimpleObject with session metadata
-     */
-    @RequestMapping(value = "/rules/{sessionId}", method = RequestMethod.GET)
-    @ResponseBody
-    public SimpleObject getSessionDetails(@PathVariable("sessionId") String sessionId) {
-        SessionMetadata metadata = sessionRegistry.getSessionMetadata(sessionId);
-        if (metadata == null) {
-            throw new ResponseStatusException(
-                    HttpStatus.NOT_FOUND,
-                    String.format("Session '%s' not found in registry", sessionId)
-            );
-        }
-
-        SimpleObject result = new SimpleObject();
-        result.add("sessionId", metadata.getSessionId());
-        result.add("autoStartable", metadata.isAutoStartable());
-        result.add("createdAt", metadata.getCreatedAt().toString());
-        result.add("createdByThread", metadata.getCreatedByThread());
-        result.add("lastAccessed", metadata.getLastAccessed().toString());
-        result.add("accessCount", metadata.getAccessCount());
-        return result;
-    }
-
-    /**
-     * POST /ws/rest/v1/drools/rules/{sessionId}/execute
-     * Executes rules against an existing session in the registry.
-     * Only works for auto-startable sessions that are already registered.
-     *
-     * @param sessionId the session identifier
-     * @param allParams execution parameters
-     * @param request HTTP request
-     * @return SimpleObject with execution results
-     */
-    @RequestMapping(value = "/rules/{sessionId}/execute", method = RequestMethod.POST)
-    @ResponseBody
-    public SimpleObject executeAgainstSession(
-            @PathVariable("sessionId") String sessionId,
-            @RequestParam Map<String, String> allParams,
-            HttpServletRequest request) {
-
-        if (!sessionRegistry.sessionExists(sessionId)) {
-            throw new ResponseStatusException(
-                    HttpStatus.NOT_FOUND,
-                    String.format("Session '%s' not found in registry. Only auto-startable sessions can be queried.", sessionId)
-            );
-        }
-
-        DroolsSessionConfig sessionConfig = droolsService.getSessionConfig(sessionId);
-        if (sessionConfig == null) {
-            throw new ResponseStatusException(
-                    HttpStatus.NOT_FOUND,
-                    String.format("Session configuration for '%s' not found", sessionId)
-            );
-        }
-
-        validateParams(sessionConfig, allParams);
-        return convertToSimpleObject(
-                sessionExecutor.executeAgainstExistingSession(sessionId, allParams),
-                request
-        );
-    }
-
-    /**
-     * DELETE /ws/rest/v1/drools/rules/{sessionId}
-     * Removes and disposes a session from the registry.
-     *
-     * @param sessionId the session identifier
-     * @return SimpleObject with success status
-     */
-    @RequestMapping(value = "/rules/{sessionId}", method = RequestMethod.DELETE)
-    @ResponseBody
-    public SimpleObject disposeSession(@PathVariable("sessionId") String sessionId) {
-        boolean removed = sessionRegistry.removeSession(sessionId);
-
-        SimpleObject result = new SimpleObject();
-        result.add("sessionId", sessionId);
-        result.add("removed", removed);
-        result.add("message", removed
-                ? String.format("Session '%s' successfully removed and disposed", sessionId)
-                : String.format("Session '%s' not found in registry", sessionId)
-        );
-        return result;
-    }
-
-    /**
-     * GET /ws/rest/v1/drools/rules/{sessionId}/exists
-     * Checks if a session exists in the registry.
-     *
-     * @param sessionId the session identifier
-     * @return SimpleObject with existence status
-     */
-    @RequestMapping(value = "/rules/{sessionId}/exists", method = RequestMethod.GET)
-    @ResponseBody
-    public SimpleObject sessionExists(@PathVariable("sessionId") String sessionId) {
-        boolean exists = sessionRegistry.sessionExists(sessionId);
-
-        SimpleObject result = new SimpleObject();
-        result.add("sessionId", sessionId);
-        result.add("exists", exists);
-        return result;
     }
 
     private void validateParams(DroolsSessionConfig sessionConfig, Map<String, String> params) {
